Stats

ArviZ.ArviZStats.AbstractELPDResult

ArviZ.ArviZStats.AbstractModelWeightsMethod

ArviZ.ArviZStats.BootstrappedPseudoBMA

ArviZ.ArviZStats.ModelComparisonResult

ArviZ.ArviZStats.PSISLOOResult

ArviZ.ArviZStats.PseudoBMA

ArviZ.ArviZStats.Stacking

ArviZ.ArviZStats.SummaryStats

ArviZ.ArviZStats.WAICResult

PSIS.PSISResult

ArviZ.ArviZStats.compare

ArviZ.ArviZStats.elpd_estimates

ArviZ.ArviZStats.hdi

ArviZ.ArviZStats.hdi!

ArviZ.ArviZStats.information_criterion

ArviZ.ArviZStats.loo

ArviZ.ArviZStats.loo_pit

ArviZ.ArviZStats.model_weights

ArviZ.ArviZStats.r2_score

ArviZ.ArviZStats.smooth_data

ArviZ.ArviZStats.waic

PSIS.psis

PSIS.psis!

StatsBase.summarystats

Summary statistics

ArviZ.ArviZStats.SummaryStats — Type

A container for a column table of values computed by summarystats.

This object implements the Tables and TableTraits interfaces and has a custom show method.

data: The summary statistics for each variable, with the first entry containing the variable names

StatsBase.summarystats — Function

summarystats(data::InferenceData; group=:posterior, kwargs...)
summarystats(data::Dataset; kwargs...)

Compute summary statistics and diagnostics on the data.

Keywords

prob_interval::Real: The value of the prob argument to hdi used to compute the highest density interval. Defaults to 0.94.

return_type::Type: The type of object to return. Valid options are Dataset and SummaryStats. Defaults to SummaryStats.

metric_dim: The dimension name or type to use for the computed metrics. Only used if return_type is Dataset. Defaults to Dim{:_metric}.

compact_labels::Bool: Whether to use compact names for the variables. Only used if return_type is SummaryStats. Defaults to true.

kind::Symbol: Whether to compute just statistics (:stats), just diagnostics (:diagnostics), or both (:both). Defaults to :both.

Examples

Compute the summary statistics and diagnostics on posterior draws of the centered eight model:

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> summarystats(idata.posterior[(:mu, :tau)])
SummaryStats
      mean  std  hdi_3%  hdi_97%  mcse_mean  mcse_std  ess_tail  ess_bulk  rha ⋯
 mu    4.5  3.5  -1.62     10.7        0.23      0.11       659       241  1.0 ⋯
 tau   4.1  3.1   0.896     9.67       0.26      0.17        38        67  1.0 ⋯
                                                                1 column omitted

Compute just the statistics on all variables:

julia> summarystats(idata.posterior; kind=:stats)
SummaryStats
                          mean   std  hdi_3%  hdi_97%
 mu                       4.49  3.49  -1.62     10.7
 theta[Choate]            6.46  5.87  -4.56     17.1
 theta[Deerfield]         5.03  4.88  -4.31     14.3
 theta[Phillips Andover]  3.94  5.69  -7.77     13.7
 theta[Phillips Exeter]   4.87  5.01  -4.49     14.7
 theta[Hotchkiss]         3.67  4.96  -6.47     11.7
 theta[Lawrenceville]     3.97  5.19  -7.04     12.2
 theta[St. Paul's]        6.58  5.11  -3.09     16.3
 theta[Mt. Hermon]        4.77  5.74  -5.86     16.0
 tau                      4.12  3.10   0.896     9.67

Compute the statistics and diagnostics from the posterior group of an InferenceData and store in a Dataset:

julia> summarystats(idata; return_type=Dataset)
Dataset with dimensions:
  Dim{:_metric} Categorical{String} String[mean, std, …, ess_bulk, rhat] Unordered,
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered
and 3 layers:
  :mu    Float64 dims: Dim{:_metric} (9)
  :theta Float64 dims: Dim{:school}, Dim{:_metric} (8×9)
  :tau   Float64 dims: Dim{:_metric} (9)

General statistics

ArviZ.ArviZStats.hdi — Function

hdi(samples::AbstractArray{<:Real}; prob=0.94) -> (; lower, upper)

Estimate the unimodal highest density interval (HDI) of samples for the probability prob.

The HDI is the minimum width Bayesian credible interval (BCI). That is, it is the smallest possible interval containing at least (100*prob)% of the draws.[Hyndman1996]

samples is an array of shape (draws[, chains[, params...]]). If multiple parameters are present, then lower and upper are arrays with the shape (params...,), computed separately for each marginal.

This implementation uses the algorithm of [ChenShao1999].

Note

Any default value of prob is arbitrary. The default value of prob=0.94 instead of a more common default like prob=0.95 is chosen to reminder the user of this arbitrariness.

Examples

Here we calculate the 83% HDI for a normal random variable:

julia> using ArviZ

julia> x = randn(2_000);

julia> hdi(x; prob=0.83) |> pairs
pairs(::NamedTuple) with 2 entries:
  :lower => -1.38266
  :upper => 1.25982

We can also calculate the HDI for a 3-dimensional array of samples:

julia> x = randn(1_000, 1, 1) .+ reshape(0:5:10, 1, 1, :);

julia> hdi(x) |> pairs
pairs(::NamedTuple) with 2 entries:
  :lower => [-1.9674, 3.0326, 8.0326]
  :upper => [1.90028, 6.90028, 11.9003]

hdi(data::InferenceData; prob=0.94) -> Dataset
hdi(data::Dataset; prob=0.94) -> Dataset

Calculate the highest density interval (HDI) for each parameter in the data.

Examples

Calculate HDI for all parameters in the posterior group of an InferenceData:

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> hdi(idata)
Dataset with dimensions:
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered,
  Dim{:hdi_bound} Categorical{Symbol} Symbol[:lower, :upper] ForwardOrdered
and 3 layers:
  :mu    Float64 dims: Dim{:hdi_bound} (2)
  :theta Float64 dims: Dim{:school}, Dim{:hdi_bound} (8×2)
  :tau   Float64 dims: Dim{:hdi_bound} (2)

We can also calculate the HDI for a subset of variables:

julia> hdi(idata.posterior[(:theta,)]).theta
8×2 DimArray{Float64,2} theta with dimensions:
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered,
  Dim{:hdi_bound} Categorical{Symbol} Symbol[:lower, :upper] ForwardOrdered
                        :lower    :upper
  "Choate"            -4.56375  17.1324
  "Deerfield"         -4.31055  14.2535
  "Phillips Andover"  -7.76922  13.6755
  "Phillips Exeter"   -4.48955  14.6635
  "Hotchkiss"         -6.46991  11.7191
  "Lawrenceville"     -7.04111  12.2087
  "St. Paul's"        -3.09262  16.2685
  "Mt. Hermon"        -5.85834  16.0143

ArviZ.ArviZStats.hdi! — Function

hdi!(samples::AbstractArray{<:Real}; prob=0.94) -> (; lower, upper)

A version of hdi that sorts samples in-place while computing the HDI.

ArviZ.ArviZStats.r2_score — Function

r2_score(y_true::AbstractVector, y_pred::AbstractVecOrMat) -> (; r2, r2_std)

$R²$ for linear Bayesian regression models.[GelmanGoodrich2019]

Arguments

y_true: Observed data of length noutputs

y_pred: Predicted data with size (ndraws[, nchains], noutputs)

Examples

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("regression1d");

julia> y_true = idata.observed_data.y;

julia> y_pred = PermutedDimsArray(idata.posterior_predictive.y, (:draw, :chain, :y_dim_0));

julia> r2_score(y_true, y_pred) |> pairs
pairs(::NamedTuple) with 2 entries:
  :r2     => 0.683197
  :r2_std => 0.0368838

r2_score(idata::InferenceData; y_name, y_pred_name) -> (; r2, r2_std)

Compute $R²$ from idata, automatically formatting the predictions to the correct shape.

Keywords

y_name: Name of observed data variable in idata.observed_data. If not provided, then the only observed data variable is used.

y_pred_name: Name of posterior predictive variable in idata.posterior_predictive. If not provided, then y_name is used.

Examples

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("regression10d");

julia> r2_score(idata) |> pairs
pairs(::NamedTuple) with 2 entries:
  :r2     => 0.998385
  :r2_std => 0.000100621

Pareto-smoothed importance sampling

PSIS.PSISResult — Type

PSISResult

Result of Pareto-smoothed importance sampling (PSIS) using psis.

Properties

log_weights: un-normalized Pareto-smoothed log weights

weights: normalized Pareto-smoothed weights (allocates a copy)

pareto_shape: Pareto $k=ξ$ shape parameter

nparams: number of parameters in log_weights

ndraws: number of draws in log_weights

nchains: number of chains in log_weights

reff: the ratio of the effective sample size of the unsmoothed importance ratios and the actual sample size.

ess: estimated effective sample size of estimate of mean using smoothed importance samples (see ess_is)

tail_length: length of the upper tail of log_weights that was smoothed

tail_dist: the generalized Pareto distribution that was fit to the tail of log_weights. Note that the tail weights are scaled to have a maximum of 1, so tail_dist * exp(maximum(log_ratios)) is the corresponding fit directly to the tail of log_ratios.

normalized::Bool:indicates whether log_weights are log-normalized along the sample dimensions.

Diagnostic

The pareto_shape parameter $k=ξ$ of the generalized Pareto distribution tail_dist can be used to diagnose reliability and convergence of estimates using the importance weights [VehtariSimpson2021].

if $k < \frac{1}{3}$, importance sampling is stable, and importance sampling (IS) and PSIS both are reliable.

if $k ≤ \frac{1}{2}$, then the importance ratio distributon has finite variance, and the central limit theorem holds. As $k$ approaches the upper bound, IS becomes less reliable, while PSIS still works well but with a higher RMSE.

if $\frac{1}{2} < k ≤ 0.7$, then the variance is infinite, and IS can behave quite poorly. However, PSIS works well in this regime.

if $0.7 < k ≤ 1$, then it quickly becomes impractical to collect enough importance weights to reliably compute estimates, and importance sampling is not recommended.

if $k > 1$, then neither the variance nor the mean of the raw importance ratios exists. The convergence rate is close to zero, and bias can be large with practical sample sizes.

See PSISPlots.paretoshapeplot for a diagnostic plot.

PSIS.psis — Function

psis(log_ratios, reff = 1.0; kwargs...) -> PSISResult
psis!(log_ratios, reff = 1.0; kwargs...) -> PSISResult

Compute Pareto smoothed importance sampling (PSIS) log weights [VehtariSimpson2021].

While psis computes smoothed log weights out-of-place, psis! smooths them in-place.

Arguments

log_ratios: an array of logarithms of importance ratios, with size (draws, [chains, [parameters...]]), where chains>1 would be used when chains are generated using Markov chain Monte Carlo.

reff::Union{Real,AbstractArray}: the ratio(s) of effective sample size of log_ratios and the actual sample size reff = ess/(draws * chains), used to account for autocorrelation, e.g. due to Markov chain Monte Carlo. If an array, it must have the size (parameters...,) to match log_ratios.

Keywords

warn=true: If true, warning messages are delivered

normalize=true: If true, the log-weights will be log-normalized so that exp.(log_weights) sums to 1 along the sample dimensions.

Returns

result: a PSISResult object containing the results of the Pareto-smoothing.

A warning is raised if the Pareto shape parameter $k ≥ 0.7$. See PSISResult for details and PSISPlots.paretoshapeplot for a diagnostic plot.

PSIS.psis! — Function

psis(log_ratios, reff = 1.0; kwargs...) -> PSISResult
psis!(log_ratios, reff = 1.0; kwargs...) -> PSISResult

Compute Pareto smoothed importance sampling (PSIS) log weights [VehtariSimpson2021].

While psis computes smoothed log weights out-of-place, psis! smooths them in-place.

Arguments

log_ratios: an array of logarithms of importance ratios, with size (draws, [chains, [parameters...]]), where chains>1 would be used when chains are generated using Markov chain Monte Carlo.

reff::Union{Real,AbstractArray}: the ratio(s) of effective sample size of log_ratios and the actual sample size reff = ess/(draws * chains), used to account for autocorrelation, e.g. due to Markov chain Monte Carlo. If an array, it must have the size (parameters...,) to match log_ratios.

Keywords

warn=true: If true, warning messages are delivered

normalize=true: If true, the log-weights will be log-normalized so that exp.(log_weights) sums to 1 along the sample dimensions.

Returns

result: a PSISResult object containing the results of the Pareto-smoothing.

A warning is raised if the Pareto shape parameter $k ≥ 0.7$. See PSISResult for details and PSISPlots.paretoshapeplot for a diagnostic plot.

LOO and WAIC

ArviZ.ArviZStats.AbstractELPDResult — Type

abstract type AbstractELPDResult

An abstract type representing the result of an ELPD computation.

Every subtype stores estimates of both the expected log predictive density (elpd) and the effective number of parameters p, as well as standard errors and pointwise estimates of each, from which other relevant estimates can be computed.

Subtypes implement the following functions:

elpd_estimates

information_criterion

ArviZ.ArviZStats.PSISLOOResult — Type

Results of Pareto-smoothed importance sampling leave-one-out cross-validation (PSIS-LOO).

See also: loo, AbstractELPDResult

estimates: Estimates of the expected log pointwise predictive density (ELPD) and effective number of parameters (p)

pointwise: Pointwise estimates

psis_result: Pareto-smoothed importance sampling (PSIS) results

ArviZ.ArviZStats.WAICResult — Type

Results of computing the widely applicable information criterion (WAIC).

See also: waic, AbstractELPDResult

estimates: Estimates of the expected log pointwise predictive density (ELPD) and effective number of parameters (p)

pointwise: Pointwise estimates

ArviZ.ArviZStats.elpd_estimates — Function

elpd_estimates(result::AbstractELPDResult; pointwise=false) -> (; elpd, elpd_mcse, lpd)

Return the (E)LPD estimates from the result.

ArviZ.ArviZStats.information_criterion — Function

information_criterion(elpd, scale::Symbol)

Compute the information criterion for the given scale from the elpd estimate.

scale must be one of (:deviance, :log, :negative_log).

See also: effective_number_of_parameters, loo, waic

information_criterion(result::AbstractELPDResult, scale::Symbol; pointwise=false)

Compute information criterion for the given scale from the existing ELPD result.

scale must be one of (:deviance, :log, :negative_log).

If pointwise=true, then pointwise estimates are returned.

ArviZ.ArviZStats.loo — Function

loo(log_likelihood; reff=nothing, kwargs...) -> PSISLOOResult{<:NamedTuple,<:NamedTuple}

Compute the Pareto-smoothed importance sampling leave-one-out cross-validation (PSIS-LOO). [Vehtari2017][LOOFAQ]

log_likelihood must be an array of log-likelihood values with shape (chains, draws[, params...]).

Keywords

reff::Union{Real,AbstractArray{<:Real}}: The relative effective sample size(s) of the likelihood values. If an array, it must have the same data dimensions as the corresponding log-likelihood variable. If not provided, then this is estimated using ess.

kwargs: Remaining keywords are forwarded to psis.

See also: PSISLOOResult, waic

Examples

Manually compute $R_\mathrm{eff}$ and calculate PSIS-LOO of a model:

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> log_like = PermutedDimsArray(idata.log_likelihood.obs, (:draw, :chain, :school));

julia> reff = ess(log_like; kind=:basic, split_chains=1, relative=true);

julia> loo(log_like; reff)
PSISLOOResult with estimates
 elpd  elpd_mcse    p  p_mcse
  -31        1.4  0.9    0.34

and PSISResult with 500 draws, 4 chains, and 8 parameters
Pareto shape (k) diagnostic values:
                    Count      Min. ESS
 (-Inf, 0.5]  good  7 (87.5%)  151
  (0.5, 0.7]  okay  1 (12.5%)  446

loo(data::Dataset; [var_name::Symbol,] kwargs...) -> PSISLOOResult{<:NamedTuple,<:Dataset}
loo(data::InferenceData; [var_name::Symbol,] kwargs...) -> PSISLOOResult{<:NamedTuple,<:Dataset}

Compute PSIS-LOO from log-likelihood values in data.

If more than one log-likelihood variable is present, then var_name must be provided.

Examples

Calculate PSIS-LOO of a model:

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> loo(idata)
PSISLOOResult with estimates
 elpd  elpd_mcse    p  p_mcse
  -31        1.4  0.9    0.34

and PSISResult with 500 draws, 4 chains, and 8 parameters
Pareto shape (k) diagnostic values:
                    Count      Min. ESS
 (-Inf, 0.5]  good  6 (75.0%)  135
  (0.5, 0.7]  okay  2 (25.0%)  421

ArviZ.ArviZStats.waic — Function

waic(log_likelihood::AbstractArray) -> WAICResult{<:NamedTuple,<:NamedTuple}

Compute the widely applicable information criterion (WAIC).[Watanabe2010][Vehtari2017][LOOFAQ]

log_likelihood must be an array of log-likelihood values with shape (chains, draws[, params...]).

See also: WAICResult, loo

Examples

Calculate WAIC of a model:

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> log_like = PermutedDimsArray(idata.log_likelihood.obs, (:draw, :chain, :school));

julia> waic(log_like)
WAICResult with estimates
 elpd  elpd_mcse    p  p_mcse
  -31        1.4  0.9    0.33

waic(data::Dataset; [var_name::Symbol]) -> WAICResult{<:NamedTuple,<:Dataset}
waic(data::InferenceData; [var_name::Symbol]) -> WAICResult{<:NamedTuple,<:Dataset}

Compute WAIC from log-likelihood values in data.

If more than one log-likelihood variable is present, then var_name must be provided.

Examples

Calculate WAIC of a model:

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> waic(idata)
WAICResult with estimates
 elpd  elpd_mcse    p  p_mcse
  -31        1.4  0.9    0.33

Model comparison

ArviZ.ArviZStats.ModelComparisonResult — Type

ModelComparisonResult

Result of model comparison using ELPD.

This struct implements the Tables and TableTraits interfaces.

Each field returns a collection of the corresponding entry for each model:

name: Names of the models, if provided.

rank: Ranks of the models (ordered by decreasing ELPD)

elpd_diff: ELPD of a model subtracted from the largest ELPD of any model

elpd_diff_mcse: Monte Carlo standard error of the ELPD difference

weight: Model weights computed with weights_method

elpd_result: AbstactELPDResults for each model, which can be used to access useful stats like ELPD estimates, pointwise estimates, and Pareto shape values for PSIS-LOO

weights_method: Method used to compute model weights with model_weights

ArviZ.ArviZStats.compare — Function

compare(models; kwargs...)

Compare models based on their expected log pointwise predictive density (ELPD).

models is a Tuple, NamedTuple, or AbstractVector whose values are either AbstractELPDResult entries or any argument to elpd_method, which must produce an AbstractELPDResult.

The weights are returned in the same type of collection.

The argument may be any object with a pairs method where each value is either an InferenceData or an AbstractELPDResult.

The ELPD is estimated either by Pareto smoothed importance sampling leave-one-out cross-validation (LOO) or using the widely applicable information criterion (WAIC). We recommend loo. Read more theory here - in a paper by some of the leading authorities on model comparison dx.doi.org/10.1111/1467-9868.00353

Arguments

models: a Tuple, NamedTuple, or AbstractVector whose values are either AbstractELPDResult entries or any argument to elpd_method.

Keywords

weights_method::AbstractModelWeightsMethod=Stacking(): the method to be used to weight the models. See model_weights for details

elpd_method=loo: a method that computes an AbstractELPDResult from an argument in models.

sort::Bool=true: Whether to sort models by decreasing ELPD.

Returns

ModelComparisonResult: A container for the model comparison results.

Examples

Compare the centered and non centered models of the eight school problem using the defaults: loo and Stacking weights:

julia> using ArviZ, ArviZExampleData

julia> models = (
           centered=load_example_data("centered_eight"),
           non_centered=load_example_data("non_centered_eight"),
       );

julia> mc = compare(models)
┌ Warning: 1 parameters had Pareto shape values 0.7 < k ≤ 1. Resulting importance sampling estimates are likely to be unstable.
└ @ PSIS ~/.julia/packages/PSIS/...
ModelComparisonResult with Stacking weights
               rank  elpd  elpd_mcse  elpd_diff  elpd_diff_mcse  weight    p   ⋯
 non_centered     1   -31        1.4       0              0.0       1.0  0.9   ⋯
 centered         2   -31        1.4       0.06           0.067     0.0  0.9   ⋯
                                                                1 column omitted

Compare the same models from pre-computed PSIS-LOO results and computing BootstrappedPseudoBMA weights:

julia> elpd_results = mc.elpd_result;

julia> compare(elpd_results; weights_method=BootstrappedPseudoBMA())
ModelComparisonResult with BootstrappedPseudoBMA weights
               rank  elpd  elpd_mcse  elpd_diff  elpd_diff_mcse  weight    p   ⋯
 non_centered     1   -31        1.4       0              0.0      0.52  0.9   ⋯
 centered         2   -31        1.4       0.06           0.067    0.48  0.9   ⋯
                                                                1 column omitted

ArviZ.ArviZStats.model_weights — Function

model_weights(elpd_results; method=Stacking())
model_weights(method::AbstractModelWeightsMethod, elpd_results)

Compute weights for each model in elpd_results using method.

elpd_results is a Tuple, NamedTuple, or AbstractVector with AbstractELPDResult entries. The weights are returned in the same type of collection.

Stacking is the recommended approach, as it performs well even when the true data generating process is not included among the candidate models. See [YaoVehtari2018] for details.

See also: AbstractModelWeightsMethod, compare

Examples

Compute Stacking weights for two models:

julia> using ArviZ, ArviZExampleData

julia> models = (
           centered=load_example_data("centered_eight"),
           non_centered=load_example_data("non_centered_eight"),
       );

julia> elpd_results = map(loo, models);
┌ Warning: 1 parameters had Pareto shape values 0.7 < k ≤ 1. Resulting importance sampling estimates are likely to be unstable.
└ @ PSIS ~/.julia/packages/PSIS/...

julia> model_weights(elpd_results; method=Stacking()) |> pairs
pairs(::NamedTuple) with 2 entries:
  :centered     => 5.34175e-19
  :non_centered => 1.0

Now we compute BootstrappedPseudoBMA weights for the same models:

julia> model_weights(elpd_results; method=BootstrappedPseudoBMA()) |> pairs
pairs(::NamedTuple) with 2 entries:
  :centered     => 0.483723
  :non_centered => 0.516277

The following model weighting methods are available

ArviZ.ArviZStats.AbstractModelWeightsMethod — Type

abstract type AbstractModelWeightsMethod

An abstract type representing methods for computing model weights.

Subtypes implement model_weights(method, elpd_results).

ArviZ.ArviZStats.BootstrappedPseudoBMA — Type

struct BootstrappedPseudoBMA{R<:Random.AbstractRNG, T<:Real} <: AbstractModelWeightsMethod

Model weighting method using pseudo Bayesian Model Averaging using Akaike-type weighting with the Bayesian bootstrap (pseudo-BMA+)[YaoVehtari2018].

The Bayesian bootstrap stabilizes the model weights.

BootstrappedPseudoBMA(; rng=Random.default_rng(), samples=1_000, alpha=1)
BootstrappedPseudoBMA(rng, samples, alpha)

Construct the method.

rng::Random.AbstractRNG: The random number generator to use for the Bayesian bootstrap

samples::Int64: The number of samples to draw for bootstrapping

alpha::Real: The shape parameter in the Dirichlet distribution used for the Bayesian bootstrap. The default (1) corresponds to a uniform distribution on the simplex.

See also: Stacking

ArviZ.ArviZStats.PseudoBMA — Type

struct PseudoBMA <: AbstractModelWeightsMethod

Model weighting method using pseudo Bayesian Model Averaging (pseudo-BMA) and Akaike-type weighting.

PseudoBMA(; regularize=false)
PseudoBMA(regularize)

Construct the method with optional regularization of the weights using the standard error of the ELPD estimate.

Note

This approach is not recommended, as it produces unstable weight estimates. It is recommended to instead use BootstrappedPseudoBMA to stabilize the weights or Stacking. For details, see [YaoVehtari2018].

See also: Stacking

ArviZ.ArviZStats.Stacking — Type

struct Stacking{O<:Optim.AbstractOptimizer} <: AbstractModelWeightsMethod

Model weighting using stacking of predictive distributions[YaoVehtari2018].

Stacking(; optimizer=Optim.LBFGS(), options=Optim.Options()
Stacking(optimizer[, options])

Construct the method, optionally customizing the optimization.

optimizer::Optim.AbstractOptimizer: The optimizer to use for the optimization of the weights. The optimizer must support projected gradient optimization viae a manifold field.

options::Optim.Options: The Optim options to use for the optimization of the weights.

See also: BootstrappedPseudoBMA

Predictive checks

ArviZ.ArviZStats.loo_pit — Function

loo_pit(y, y_pred, log_weights; kwargs...) -> Union{Real,AbstractArray}

Compute leave-one-out probability integral transform (LOO-PIT) checks.

Arguments

y: array of observations with shape (params...,)

y_pred: array of posterior predictive samples with shape (draws, chains, params...).

log_weights: array of normalized log LOO importance weights with shape (draws, chains, params...).

Keywords

is_discrete: If not provided, then it is set to true iff elements of y and y_pred are all integer-valued. If true, then data are smoothed using smooth_data to make them non-discrete before estimating LOO-PIT values.

kwargs: Remaining keywords are forwarded to smooth_data if data is discrete.

Returns

pitvals: LOO-PIT values with same size as y. If y is a scalar, then pitvals is a scalar.

LOO-PIT is a marginal posterior predictive check. If $y_{-i}$ is the array $y$ of observations with the $i$th observation left out, and $y_i^*$ is a posterior prediction of the $i$th observation, then the LOO-PIT value for the $i$th observation is defined as

\[P(y_i^* \le y_i \mid y_{-i}) = \int_{-\infty}^{y_i} p(y_i^* \mid y_{-i}) \mathrm{d} y_i^*\]

The LOO posterior predictions and the corresponding observations should have similar distributions, so if conditional predictive distributions are well-calibrated, then all LOO-PIT values should be approximately uniformly distributed on $[0, 1]$.[Gabry2019]

Examples

Calculate LOO-PIT values using as test quantity the observed values themselves.

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> log_weights = loo(idata; var_name=:obs).psis_result.log_weights;

julia> loo_pit(
            idata.observed_data.obs,
            permutedims(idata.posterior_predictive.obs, (:draw, :chain, :school)),
            log_weights,
        )
8-element DimArray{Float64,1} with dimensions:
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered
 "Choate"            0.943511
 "Deerfield"         0.63797
 "Phillips Andover"  0.316697
 "Phillips Exeter"   0.582252
 "Hotchkiss"         0.295321
 "Lawrenceville"     0.403318
 "St. Paul's"        0.902508
 "Mt. Hermon"        0.655275

Calculate LOO-PIT values using as test quantity the square of the difference between each observation and mu.

julia> using DimensionalData, Statistics

julia> T = idata.observed_data.obs .- only(median(idata.posterior.mu; dims=(:draw, :chain)));

julia> T_pred = permutedims(
           broadcast_dims(-, idata.posterior_predictive.obs, idata.posterior.mu),
           (:draw, :chain, :school),
       );

julia> loo_pit(T .^ 2, T_pred .^ 2, log_weights)
8-element DimArray{Float64,1} with dimensions:
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered
 "Choate"            0.873577
 "Deerfield"         0.243686
 "Phillips Andover"  0.357563
 "Phillips Exeter"   0.149908
 "Hotchkiss"         0.435094
 "Lawrenceville"     0.220627
 "St. Paul's"        0.775086
 "Mt. Hermon"        0.296706

loo_pit(idata::InferenceData, log_weights; kwargs...) -> DimArray

Compute LOO-PIT values using existing normalized log LOO importance weights.

Keywords

y_name: Name of observed data variable in idata.observed_data. If not provided, then the only observed data variable is used.

y_pred_name: Name of posterior predictive variable in idata.posterior_predictive. If not provided, then y_name is used.

kwargs: Remaining keywords are forwarded to loo_pit.

Examples

Calculate LOO-PIT values using already computed log weights.

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> loo_result = loo(idata; var_name=:obs);

julia> loo_pit(idata, loo_result.psis_result.log_weights; y_name=:obs)
8-element DimArray{Float64,1} loo_pit_obs with dimensions:
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered
 "Choate"            0.943511
 "Deerfield"         0.63797
 "Phillips Andover"  0.316697
 "Phillips Exeter"   0.582252
 "Hotchkiss"         0.295321
 "Lawrenceville"     0.403318
 "St. Paul's"        0.902508
 "Mt. Hermon"        0.655275

loo_pit(idata::InferenceData; kwargs...) -> DimArray

Compute LOO-PIT from groups in idata using PSIS-LOO.

See also: loo, psis

Keywords

y_name: Name of observed data variable in idata.observed_data. If not provided, then the only observed data variable is used.

y_pred_name: Name of posterior predictive variable in idata.posterior_predictive. If not provided, then y_name is used.

log_likelihood_name: Name of log-likelihood variable in idata.log_likelihood. If not provided, then y_name is used if idata has a log_likelihood group, otherwise the only variable is used.

reff::Union{Real,AbstractArray{<:Real}}: The relative effective sample size(s) of the likelihood values. If an array, it must have the same data dimensions as the corresponding log-likelihood variable. If not provided, then this is estimated using ess.

kwargs: Remaining keywords are forwarded to loo_pit.

Examples

Calculate LOO-PIT values using as test quantity the observed values themselves.

julia> using ArviZ, ArviZExampleData

julia> idata = load_example_data("centered_eight");

julia> loo_pit(idata; y_name=:obs)
8-element DimArray{Float64,1} loo_pit_obs with dimensions:
  Dim{:school} Categorical{String} String[Choate, Deerfield, …, St. Paul's, Mt. Hermon] Unordered
 "Choate"            0.943511
 "Deerfield"         0.63797
 "Phillips Andover"  0.316697
 "Phillips Exeter"   0.582252
 "Hotchkiss"         0.295321
 "Lawrenceville"     0.403318
 "St. Paul's"        0.902508
 "Mt. Hermon"        0.655275

Utilities

ArviZ.ArviZStats.smooth_data — Function

smooth_data(y; dims=:, interp_method=CubicSpline, offset_frac=0.01)

Smooth y along dims using interp_method.

interp_method is a 2-argument callabale that takes the arguments y and x and returns a DataInterpolations.jl interpolation method, defaulting to a cubic spline interpolator.

offset_frac is the fraction of the length of y to use as an offset when interpolating.

Hyndman1996Rob J. Hyndman (1996) Computing and Graphing Highest Density Regions, Amer. Stat., 50(2): 120-6. DOI: 10.1080/00031305.1996.10474359 jstor.

ChenShao1999Ming-Hui Chen & Qi-Man Shao (1999) Monte Carlo Estimation of Bayesian Credible and HPD Intervals, J Comput. Graph. Stat., 8:1, 69-92. DOI: 10.1080/10618600.1999.10474802 jstor.

GelmanGoodrich2019Andrew Gelman, Ben Goodrich, Jonah Gabry & Aki Vehtari (2019) R-squared for Bayesian Regression Models, The American Statistician, 73:3, 307-9, DOI: 10.1080/00031305.2018.1549100.

VehtariSimpson2021Vehtari A, Simpson D, Gelman A, Yao Y, Gabry J. (2021). Pareto smoothed importance sampling. arXiv:1507.02646v7 [stat.CO]

VehtariSimpson2021Vehtari A, Simpson D, Gelman A, Yao Y, Gabry J. (2021). Pareto smoothed importance sampling. arXiv:1507.02646v7 [stat.CO]

VehtariSimpson2021Vehtari A, Simpson D, Gelman A, Yao Y, Gabry J. (2021). Pareto smoothed importance sampling. arXiv:1507.02646v7 [stat.CO]

Vehtari2017Vehtari, A., Gelman, A. & Gabry, J. Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. Stat Comput 27, 1413–1432 (2017). doi: 10.1007/s11222-016-9696-4 arXiv: 1507.04544

LOOFAQAki Vehtari. Cross-validation FAQ. https://mc-stan.org/loo/articles/online-only/faq.html

Watanabe2010Watanabe, S. Asymptotic Equivalence of Bayes Cross Validation and Widely Applicable Information Criterion in Singular Learning Theory. 11(116):3571−3594, 2010. https://jmlr.csail.mit.edu/papers/v11/watanabe10a.html

Vehtari2017Vehtari, A., Gelman, A. & Gabry, J. Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. Stat Comput 27, 1413–1432 (2017). doi: 10.1007/s11222-016-9696-4 arXiv: 1507.04544

LOOFAQAki Vehtari. Cross-validation FAQ. https://mc-stan.org/loo/articles/online-only/faq.html

YaoVehtari2018
Yuling Yao, Aki Vehtari, Daniel Simpson, and Andrew Gelman. Using Stacking to Average Bayesian Predictive Distributions.

Bayesian Analysis. 13, 3, 917–1007.

doi: 10.1214/17-BA1091 arXiv: 1704.02030

YaoVehtari2018
Yuling Yao, Aki Vehtari, Daniel Simpson, and Andrew Gelman. Using Stacking to Average Bayesian Predictive Distributions.

Bayesian Analysis. 13, 3, 917–1007.

doi: 10.1214/17-BA1091 arXiv: 1704.02030

YaoVehtari2018
Yuling Yao, Aki Vehtari, Daniel Simpson, and Andrew Gelman. Using Stacking to Average Bayesian Predictive Distributions.

Bayesian Analysis. 13, 3, 917–1007.

doi: 10.1214/17-BA1091 arXiv: 1704.02030

YaoVehtari2018
Yuling Yao, Aki Vehtari, Daniel Simpson, and Andrew Gelman. Using Stacking to Average Bayesian Predictive Distributions.

Bayesian Analysis. 13, 3, 917–1007.

doi: 10.1214/17-BA1091 arXiv: 1704.02030

Gabry2019Gabry, J., Simpson, D., Vehtari, A., Betancourt, M. & Gelman, A. Visualization in Bayesian Workflow. J. R. Stat. Soc. Ser. A Stat. Soc. 182, 389–402 (2019). doi: 10.1111/rssa.12378 arXiv: 1709.01449