InferenceData

ArviZ.InferenceData

Base.getindex

Base.getproperty

Base.propertynames

Base.setindex

Type definition

ArviZ.InferenceData — Type

InferenceData{group_names,group_types}

Container for inference data storage using DimensionalData.

This object implements the InferenceData schema.

Internally, groups are stored in a NamedTuple, which can be accessed using parent(::InferenceData).

Constructors

InferenceData(groups::NamedTuple)
InferenceData(; groups...)

Construct an inference data from either a NamedTuple or keyword arguments of groups.

Groups must be Dataset objects.

Instead of directly creating an InferenceData, use the exported from_xyz functions or convert_to_inference_data.

Property interface

Base.getproperty — Function

getproperty(data::InferenceData, name::Symbol) -> Dataset

Get group with the specified name.

Base.propertynames — Function

propertynames(data::InferenceData) -> Tuple{Symbol}

Get names of groups

Indexing interface

Base.getindex — Function

Base.getindex(data::InferenceData, groups::Symbol; coords...) -> Dataset
Base.getindex(data::InferenceData, groups; coords...) -> InferenceData

Return a new InferenceData containing the specified groups sliced to the specified coords.

coords specifies a dimension name mapping to an index, a DimensionalData.Selector, or an IntervalSets.AbstractInterval.

If one or more groups lack the specified dimension, a warning is raised but can be ignored. All groups that contain the dimension must also contain the specified indices, or an exception will be raised.

Examples

Select data from all groups for just the specified schools.

julia> using ArviZ, DimensionalData

julia> idata = load_arviz_data("centered_eight");

julia> idata_sel = idata[school=At(["Choate", "Deerfield"])]
InferenceData with groups:
  > posterior
  > posterior_predictive
  > sample_stats
  > prior
  > observed_data

julia> idata_sel.posterior
Dataset with dimensions:
  Dim{:chain} Sampled 0:3 ForwardOrdered Regular Points,
  Dim{:draw} Sampled 0:499 ForwardOrdered Regular Points,
  Dim{:school} Categorical String[Choate, Deerfield] Unordered
and 3 layers:
  :mu    Float64 dims: Dim{:chain}, Dim{:draw} (4×500)
  :theta Float64 dims: Dim{:chain}, Dim{:draw}, Dim{:school} (4×500×2)
  :tau   Float64 dims: Dim{:chain}, Dim{:draw} (4×500)

with metadata OrderedCollections.OrderedDict{Symbol, Any} with 3 entries:
  :created_at                => "2019-06-21T17:36:34.398087"
  :inference_library_version => "3.7"
  :inference_library         => "pymc3"

Select data from just the posterior, returning a Dataset if the indices index more than one element from any of the variables:

julia> idata[:observed_data, school=At(["Choate"])]
Dataset with dimensions:
  Dim{:school} Categorical String[Choate] Unordered
and 1 layer:
  :obs Float64 dims: Dim{:school} (1)

with metadata OrderedCollections.OrderedDict{Symbol, Any} with 3 entries:
  :created_at                => "2019-06-21T17:36:34.491909"
  :inference_library_version => "3.7"
  :inference_library         => "pymc3"

Note that if a single index is provided, the behavior is still to slice so that the dimension is preserved.

Base.setindex — Function

Base.setindex(data::InferenceData, group::Dataset, name::Symbol) -> InferenceData

Create a new InferenceData containing the group with the specified name.

If a group with name is already in data, it is replaced.

Iteration interface

InferenceData also implements the same iteration interface as its underlying NamedTuple. That is, iterating over an InferenceData iterates over its groups.