API Reference

ArviZ.InferenceData — Type

InferenceData(::PyObject)
InferenceData(; kwargs...)

Loose wrapper around arviz.InferenceData, which is a container for inference data storage using xarray.

InferenceData can be constructed either from an arviz.InferenceData or from multiple Datasets assigned to groups specified as kwargs.

Instead of directly creating an InferenceData, use the exported from_xyz functions or convert_to_inference_data.

ArviZ.autocorr — Method

See documentation for arviz.autocorr.

ArviZ.autocov — Method

See documentation for arviz.autocov.

ArviZ.bfmi — Method

See documentation for arviz.bfmi.

ArviZ.compare — Method

See documentation for arviz.compare.

ArviZ.concat! — Method

concat!(data::InferenceData, args::InferenceData...; kwargs...) -> InferenceData

In-place version of concat, where data is modified to contain the concatenation of data and args. See concat for a description of kwargs.

ArviZ.concat — Method

See documentation for arviz.concat.

ArviZ.convert_to_inference_data — Method

See documentation for arviz.convert_to_inference_data.

ArviZ.convert_to_inference_data — Method

convert_to_inference_data(obj::AbstractChains; group = :posterior, kwargs...) -> InferenceData

Convert the chains obj to an InferenceData with the specified group. Remaining kwargs are forwarded to from_mcmcchains.

ArviZ.ess — Method

See documentation for arviz.ess.

ArviZ.from_cmdstan — Method

See documentation for arviz.from_cmdstan.

ArviZ.from_cmdstan — Method

from_cmdstan(posterior::AbstractChains; kwargs...) -> InferenceData

Call from_mcmcchains on output of CmdStan.

ArviZ.from_dict — Method

See documentation for arviz.from_dict.

ArviZ.from_mcmcchains — Function

from_mcmcchains(posterior::AbstractChains; kwargs...) -> InferenceData
from_mcmcchains(; kwargs...) -> InferenceData

Convert data in an MCMCChains.AbstractChains format into an InferenceData.

Any keyword argument below without an an explicitly annotated type above is allowed, so long as it can be passed to convert_to_dataset.

Arguments

posterior::AbstractChains: Draws from the posterior

Keywords

posterior_predictive=nothing: Draws from the posterior predictive distribution or name(s) of predictive variables in posterior

prior::AbstractChains=nothing: Draws from the prior

prior_predictive=nothing: Draws from the prior predictive distribution or name(s) of predictive variables in prior

observed_data=nothing: Data actually observed or name(s) of observed data variables in posterior

constant_data=nothing: Data that are constant, not observed (e.g. metadata), or name(s) of variables in posterior

log_likelihood::String=nothing: Name of variable in posterior with log likelihoods

library=MCMCChains: Name of library that generated the chains

coords::Dict{String,Vector}=nothing: Map from named dimension to named indices

dims::Dict{String,Vector{String}}=nothing: Map from variable name to names of its dimensions

Returns

InferenceData: The data with groups corresponding to the provided data

ArviZ.from_netcdf — Method

See documentation for arviz.from_netcdf.

ArviZ.geweke — Method

See documentation for arviz.geweke.

ArviZ.hpd — Method

See documentation for arviz.hpd.

ArviZ.load_arviz_data — Method

See documentation for arviz.load_arviz_data.

ArviZ.loo — Method

See documentation for arviz.loo.

ArviZ.loo_pit — Method

See documentation for arviz.loo_pit.

ArviZ.make_ufunc — Method

See documentation for arviz.make_ufunc.

ArviZ.mcse — Method

See documentation for arviz.mcse.

ArviZ.plot_autocorr — Method

See documentation for arviz.plot_autocorr.

ArviZ.plot_compare — Method

See documentation for arviz.plot_compare.

ArviZ.plot_density — Method

See documentation for arviz.plot_density.

ArviZ.plot_dist — Method

See documentation for arviz.plot_dist.

ArviZ.plot_elpd — Method

See documentation for arviz.plot_elpd.

ArviZ.plot_energy — Method

See documentation for arviz.plot_energy.

ArviZ.plot_ess — Method

See documentation for arviz.plot_ess.

ArviZ.plot_forest — Method

See documentation for arviz.plot_forest.

ArviZ.plot_hpd — Method

See documentation for arviz.plot_hpd.

ArviZ.plot_joint — Method

See documentation for arviz.plot_joint.

ArviZ.plot_kde — Method

See documentation for arviz.plot_kde.

ArviZ.plot_khat — Method

See documentation for arviz.plot_khat.

ArviZ.plot_loo_pit — Method

See documentation for arviz.plot_loo_pit.

ArviZ.plot_mcse — Method

See documentation for arviz.plot_mcse.

ArviZ.plot_pair — Method

See documentation for arviz.plot_pair.

ArviZ.plot_parallel — Method

See documentation for arviz.plot_parallel.

ArviZ.plot_posterior — Method

See documentation for arviz.plot_posterior.

ArviZ.plot_ppc — Method

See documentation for arviz.plot_ppc.

ArviZ.plot_rank — Method

See documentation for arviz.plot_rank.

ArviZ.plot_trace — Method

See documentation for arviz.plot_trace.

ArviZ.plot_violin — Method

See documentation for arviz.plot_violin.

ArviZ.psislw — Method

See documentation for arviz.psislw.

ArviZ.r2_score — Method

See documentation for arviz.r2_score.

ArviZ.rhat — Method

See documentation for arviz.rhat.

ArviZ.to_netcdf — Method

See documentation for arviz.to_netcdf.

ArviZ.waic — Method

See documentation for arviz.waic.

ArviZ.with_interactive_backend — Method

with_interactive_backend(f; backend::Symbol = nothing)

Execute the thunk f in a temporary interactive context with the chosen backend, or provide no arguments to use a default.

Example

idata = load_arviz_data("centered_eight")
plot_posterior(idata) # inline
with_interactive_backend() do
    plot_density(idata) # interactive
end
plot_trace(idata) # inline

ArviZ.with_rc_context — Method

with_rc_context(f; rc = nothing, fname = nothing)

Execute the thunk f within a context controlled by rc params. To see supported params, execute rc_params().

This allows one to do:

with_rc_context(fname = "pystan.rc") do
    idata = load_arviz_data("radon")
    plot_posterior(idata; var_names=["gamma"])
end

The plot would have settings from pystan.rc.

A dictionary can also be passed to the context manager:

with_rc_context(rc = Dict("plot.max_subplots" => 1), fname = "pystan.rc") do
    idata = load_arviz_data("radon")
    plot_posterior(idata, var_names=["gamma"])
end

The rc dictionary takes precedence over the settings loaded from fname. Passing a dictionary only is also valid.

ArviZ.wrap_xarray_ufunc — Method

See documentation for arviz.wrap_xarray_ufunc.

StatsBase.summarystats — Method

summarystats(data::Dataset; kwargs...) -> Union{Pandas.DataFrame,Dataset}
summarystats(data::InferenceData; group = :posterior kwargs...) -> Union{Pandas.DataFrame,Dataset}

Compute summary statistics on data.

Arguments

data::Union{Dataset,InferenceData}: The data on which to compute summary statistics. If data is an InferenceData, only the dataset corresponding to group is used.

Keywords

var_names::Vector{String}=nothing: Names of variables to include in summary

include_circ::Bool=false: Whether to include circular statistics

fmt::String="wide": Return format is either Pandas.DataFrame ("wide", "long") or Dataset ("xarray").

round_to::Int=nothing: Number of decimals used to round results. Use nothing to return raw numbers.

stat_funcs::Union{Dict{String,Function},Vector{Function}}=nothing: A vector of functions or a dict of functions with function names as keys used to calculate statistics. By default, the mean, standard deviation, simulation standard error, and highest posterior density intervals are included. The functions will be given one argument, the samples for a variable as an array, The functions should operate on an array, returning a single number. For example, Statistics.mean, or Statistics.var would both work.

extend::Bool=true: If true, use the statistics returned by stat_funcs in addition to, rather than in place of, the default statistics. This is only meaningful when stat_funcs is not nothing.

credible_interval::Real=0.94: Credible interval to plot. This is only meaningful when stat_funcs is nothing.

order::String="C": If fmt is "wide", use either "C" or "F" unpacking order.

index_origin::Int=1: If fmt is "wide", select 𝑛-based indexing for multivariate parameters.

Returns

Union{Pandas.DataFrame,Dataset}: Return type dicated by fmt argument. Return value will contain summary statistics for each variable. Default statistics are:

mean

sd

hpd_3%

hpd_97%

mcse_mean

mcse_sd

ess_bulk

ess_tail

r_hat (only computed for traces with 2 or more chains)

Examples

using ArviZ
idata = load_arviz_data("centered_eight")
summarystats(idata; var_names=["mu", "tau"])

Other statistics can be calculated by passing a list of functions or a dictionary with key, function pairs:

using StatsBase, Statistics
function median_sd(x)
    med = median(x)
    sd = sqrt(mean((x .- med).^2))
    return sd
end

func_dict = Dict(
    "std" => x -> std(x; corrected = false),
    "median_std" => median_sd,
    "5%" => x -> percentile(x, 5),
    "median" => median,
    "95%" => x -> percentile(x, 95),
)

summarystats(idata; var_names = ["mu", "tau"], stat_funcs = func_dict, extend = false)

ArviZ.Dataset — Type

Dataset(::PyObject)

Loose wrapper around xarray.Dataset, mostly used for dispatch.

To create a Dataset, use convert_to_dataset.

ArviZ.convert_to_dataset — Method

See documentation for arviz.convert_to_dataset.

ArviZ.convert_to_dataset — Method

convert_to_dataset(chns::AbstractChains; library = MCMCChains, kwargs...) -> Dataset

Convert the chains obj to a Dataset. library is the library that created the chains. Remaining kwargs are forwarded to dict_to_dataset.

ArviZ.dataset_to_dict — Method

dataset_to_dict(ds::Dataset) -> Tuple{Dict{String,Array},NamedTuple}

Convert a Dataset to a dictionary of Arrays. The function also returns keyword arguments to dict_to_dataset.

ArviZ.dict_to_dataset — Method

See documentation for arviz.dict_to_dataset.

ArviZ.from_cmdstanpy — Method

See documentation for arviz.from_cmdstanpy.

ArviZ.from_emcee — Method

See documentation for arviz.from_emcee.

ArviZ.from_numpyro — Method

See documentation for arviz.from_numpyro.

ArviZ.from_pymc3 — Method

See documentation for arviz.from_pymc3.

ArviZ.from_pyro — Method

See documentation for arviz.from_pyro.

ArviZ.from_pystan — Method

See documentation for arviz.from_pystan.

ArviZ.from_tfp — Method

See documentation for arviz.from_tfp.

ArviZ.rc_params — Method

rc_params() > Dict{String,Any}

Get the list of customizable rc params using with_rc_context.

ArviZ.replacemissing — Method

replacemissing(x)

Replace missing values with NaN and do type inference on the result.

ArviZ.reshape_values — Method

reshape_values(x::AbstractArray) -> AbstractArray

Convert from MCMCChains variable values with dimensions (ndraw, size..., nchain) to ArviZ's expected (nchain, ndraw, size...).

ArviZ.styles — Method

styles() -> Vector{String}

Get all available matplotlib styles for use with use_style

ArviZ.summary — Method

summary(
    data;
    group = :posterior,
    coords = nothing,
    dims = nothing,
    kwargs...,
) -> Union{Pandas.DataFrame,PyObject}

Compute summary statistics on any object that can be passed to convert_to_dataset.

Keywords

coords::Dict{String,Vector}=nothing: Map from named dimension to named indices.

dims::Dict{String,Vector{String}}=nothing: Map from variable name to names of its dimensions.

kwargs: Keyword arguments passed to summarystats.

ArviZ.topandas — Method

topandas(df::DataFrames.DataFrame) -> Pandas.DataFrame
topandas(df::MCMCChains.ChainDataFrame) -> Pandas.DataFrame

Convert df into a Pandas format, maintaining column order and replacing missing with NaN.

ArviZ.use_style — Method

use_style(style::String)
use_style(style::Vector{String})

Use matplotlib style settings from a style specification style.

The style name of "default" is reserved for reverting back to the default style settings.

ArviZ-specific styles are ["arviz-whitegrid", "arviz-darkgrid", "arviz-colors", "arviz-white"]. To see all available style specifications, use styles().

If a Vector of styles is provided, they are applied from first to last.

ArviZ.@forwardfun — Macro

@forwardfun f
@forwardfun(f)

Wrap a function arviz.f in f, forwarding its docstrings.